#2089 move LatexTemplate files
