+++ eo-maven-plugin/src/main/java/org/eolang/maven/LatexTemplate.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.maven.latex;
+package org.eolang.maven;
 
 import java.util.regex.Pattern;
 import org.cactoos.io.ResourceOf;
