+++ eo-maven-plugin/src/test/java/org/eolang/maven/LatexTemplateTest.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-package org.eolang.maven.latex;
+package org.eolang.maven;
 
 import org.hamcrest.MatcherAssert;
 import org.hamcrest.Matchers;
